# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a Todo CRUD application.",
        "title": "Todo API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Raphael Olaiyapo",
            "url": "https://archdemy.netlify.app/",
            "email": "raphael.tunji@yahoo.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/BillCategories": {
            "post": {
                "description": "Get bill categories.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Flutterwave"
                ],
                "summary": "Bill categories",
                "parameters": [
                    {
                        "description": "Bill Categories",
                        "name": "Todo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Model.GetBillsCatPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/Todos": {
            "get": {
                "description": "Get a list of all todos.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Todos"
                ],
                "summary": "Get a list of todos",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Model.Todo"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Add a todo to list of todos.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Todos"
                ],
                "summary": "Add a todo",
                "parameters": [
                    {
                        "description": "Add Todo",
                        "name": "Todo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Model.AddTodoDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Model.Todo"
                        }
                    }
                }
            }
        },
        "/Todos/{id}": {
            "get": {
                "description": "Get a todo from list of todos.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Todos"
                ],
                "summary": "Get a todo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Todo ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Model.Todo"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a todo from the list of todos.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Todos"
                ],
                "summary": "Delete a todo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Todo ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "Model.AddTodoDto": {
            "type": "object",
            "required": [
                "item",
                "owner"
            ],
            "properties": {
                "completed": {
                    "type": "boolean"
                },
                "item": {
                    "description": "custom validation",
                    "type": "string"
                },
                "owner": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                }
            }
        },
        "Model.GetBillsCatPayload": {
            "type": "object",
            "properties": {
                "index": {
                    "type": "string"
                },
                "queryParam": {
                    "type": "string"
                }
            }
        },
        "Model.Todo": {
            "type": "object",
            "required": [
                "item",
                "owner"
            ],
            "properties": {
                "completed": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "item": {
                    "description": "custom validation",
                    "type": "string"
                },
                "owner": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
    "externalDocs": {
        "description": "OpenAPI",
        "url": "https://swagger.io/resources/open-api/"
    }
}